Design practicum - Dictionary/Translator App
This is my first project for my program design/implementation practicum.

These 2 simple executable jars are an exercise in getting two separate
services to transmit data to each other, and then parsing that data for
use in each program.

Dictionary.jar has a GUI for ease-of-use, but Translator.jar can be run
from the cmd line as the attached documentation explains.