CMSC 256 - Data Structures/OOP final project
This project is an exercise in recursion, stacks, and iterators.

After being executed from the command line along with an argument
-maze.data, Proj73037.java will print a copy of the maze to the console
and prompt the user for valid starting coordinates.

The maze will then attempt to solve itself, and if no path is found to
the exit, will alert the user that it's "trapped."

I received a perfect score for this project, along with extra points for
having the maze print out a correct path to the exit if one is found.